Flytt aldersutregning til utils
