+++ src/skolepenger/Forside.tsx
@@ -18,6 +18,8 @@ import { hentPath } from '../utils/routing';
 import { logSidevisningSkolepenger } from '../utils/amplitude';
 import { useMount } from '../utils/hooks';
 import { ESkjemanavn } from '../utils/skjemanavn';
+import { FnrOgDnrTilAlder } from '../overgangsstønad/utils';
+import { AlertStripeFeil } from 'nav-frontend-alertstriper';
 
 const Forside: React.FC<any> = ({ intl }) => {
   const { person } = usePersonContext();
@@ -45,6 +47,8 @@ const Forside: React.FC<any> = ({ intl }) => {
     });
   };
 
+  const alder = FnrOgDnrTilAlder(person.søker.fnr);
+
   const forside = useForsideInnhold(ForsideType.skolepenger);
 
   const kanBrukeMellomlagretSøknad =
@@ -69,6 +73,14 @@ const Forside: React.FC<any> = ({ intl }) => {
             />
           </div>
 
+          {alder < 18 && (
+            <div className="ie-feil">
+              <AlertStripeFeil>
+                <LocaleTekst tekst={'side.alert.ikkeGammelNok'} />
+              </AlertStripeFeil>
+            </div>
+          )}
+
           <Sidetittel>
             <LocaleTekst tekst={'skolepenger.overskrift'} />
           </Sidetittel>
@@ -81,17 +93,19 @@ const Forside: React.FC<any> = ({ intl }) => {
               skjemanavn={ESkjemanavn.Skolepenger}
             />
           ) : (
-            <Forsideinformasjon
-              seksjon={seksjon}
-              disclaimer={disclaimer}
-              person={person}
-              intl={intl}
-              harBekreftet={søknad.harBekreftet}
-              settBekreftelse={settBekreftelse}
-              nesteSide={
-                hentPath(RoutesSkolepenger, ERouteSkolepenger.OmDeg) || ''
-              }
-            />
+            alder > 17 && (
+              <Forsideinformasjon
+                seksjon={seksjon}
+                disclaimer={disclaimer}
+                person={person}
+                intl={intl}
+                harBekreftet={søknad.harBekreftet}
+                settBekreftelse={settBekreftelse}
+                nesteSide={
+                  hentPath(RoutesSkolepenger, ERouteSkolepenger.OmDeg) || ''
+                }
+              />
+            )
           )}
         </Panel>
       </div>
